// src/components/features/daos/view/tabs/overview-tab/schema.tsx
export type CreateGateData = {
  id?: string;
  daoId?: string;
  categories: string[];
  expire_date?: string;
  claim_limit?: number;
  data_model_id?: string;
  loyalty_id?: string;
  points?: number;
  schema?: any;
  claim?: any;
} & Required<Pick<Gates, 'title' | 'categories' | 'image' | 'description'>> &
  Required<{ creator: Pick<Gates['creator'], 'id'> }> & {
    type: 'task_based' | 'direct';
    whitelisted_wallets_file?: Partial<Files>;
    isFalid?: boolean;
    tasks?: Task[];
  };

export type GateType = CreateGateData['type'];

// Task
export type SelfVerifyTask = {
  task_type: 'self_verify';
  task_data: FileTaskData;
};

export type ManualTask = {
  task_type: 'manual';
  task_data: ManualTaskData;
};

export type MeetingCodeTask = {
  task_type: 'meeting_code';
  task_data: VerificationCodeData;
};

export type QuizTask = {
  task_type: 'quiz';
  task_data: QuizTaskData;
};

export type SnapshotTask = {
  task_type: 'snapshot';
  task_data: SnapshotData;
};

export type HoldTokenTask = {
  task_type: 'token_hold';
  task_data: HoldTokenData;
};

export type HoldNFTTask = {
  task_type: 'nft_hold';
  task_data: HoldNFTData;
};

export type FollowProfileTask = {
  task_type: 'twitter_follow';
  task_data: TwitterFollowData;
};

export type TwitterTweetTask = {
  task_type: 'twitter_tweet';
  task_data: TwitterTweetData;
};

export type TwitterRetweetTask = {
  task_type: 'twitter_retweet';
  task_data: TwitterRetweetData;
};

export type TwitterLikeTask = {
  task_type: 'twitter_like';
  task_data: TwitterLikeData;
};

export type GithubContributeTask = {
  task_type: 'github_contribute';
  task_data: GithubContributeData;
};

export type GithubPRTask = {
  task_type: 'github_prs';
  task_data: GithubPRData;
};

export type Task = {
  id?: string;
  gate_id?: string;
  task_id?: string;
  title: string;
  description: string;
  order: number;
} & (
  | SelfVerifyTask
  | MeetingCodeTask
  | QuizTask
  | SnapshotTask
  | HoldTokenTask
  | HoldNFTTask
  | FollowProfileTask
  | TwitterTweetTask
  | TwitterLikeTask
  | TwitterRetweetTask
  | GithubContributeTask
  | GithubPRTask
  | ManualTask
);